// packages/spreadsheet-component/src/types/xlsx.ts
export type XMLAttributeValue = string | number | boolean;
type XMLAttribute = [string, XMLAttributeValue];
export type XMLAttributes = XMLAttribute[];

/**
 * Represent a raw XML string
 */
export class XMLString {
  /**
   * @param xmlString should be a well formed, properly escaped XML string
   */
  constructor(private xmlString: string) {}

  toString(): string {
    return this.xmlString;
  }
}

export interface XLSXDxf {
  font?: XLSXFont;
  fill?: XLSXFill;
  numFmt?: XLSXNumFormat;
  alignment?: XLSXCellAlignment;
  border?: XLSXBorder;
}
// ...
export interface XLSXFont {
  size?: number;
  family?: number;
  color?: XLSXColor;
  name?: string;
  bold?: boolean;
  italic?: boolean;
  underline?: boolean;
  strike?: boolean;
}
export interface XLSXCellAlignment {
  horizontal?: XLSXHorizontalAlignment;
  vertical?: XLSXVerticalAlignment;
  textRotation?: number;
  wrapText?: boolean;
  indent?: number;
  relativeIndent?: number;
  justifyLastLine?: boolean;
  shrinkToFit?: boolean;
  readingOrder?: number;
}

export interface XLSXFill {
  patternType?: XLSXFillPatternType;
  reservedAttribute?: string; // will generate a specific specific attribute in XML. If set, fgColor is ignored.
  fgColor?: XLSXColor;
  bgColor?: XLSXColor;
}

export interface XLSXStyle {
  fontId: number;
  fillId: number;
  borderId: number;
  numFmtId: number;
  alignment?: XLSXCellAlignment;
}
// ...
export interface XLSXNumFormat {
  id: number;
  format: Format;
}
export interface XLSXBorder {
  top?: XLSXBorderDescr;
  left?: XLSXBorderDescr;
  bottom?: XLSXBorderDescr;
  right?: XLSXBorderDescr;
  diagonal?: XLSXBorderDescr;
  diagonalUp?: boolean;
  diagonalDown?: boolean;
}

export interface XLSXBorderDescr {
  style: XLSXBorderStyle;
  color: XLSXColor;
}
// ...
type XLSXFillPatternType =
  | "none"
  | "solid"
  | "gray0625"
  | "gray125"
  | "lightGray"
  | "mediumGray"
  | "darkGray"
  | "darkHorizontal"
  | "darkVertical"
  | "darkUp"
  | "darkDown"
  | "darkGrid"
  | "darkTrellis"
  | "lightHorizontal"
  | "lightVertical"
  | "lightDown"
  | "lightUp"
  | "lightGrid"
  | "lightTrellis";

export type XLSXBorderStyle =
  | "dashDot"
  | "dashDotDot"
  | "dashed"
  | "dotted"
  | "double"
  | "hair"
  | "medium"
  | "mediumDashDot"
  | "mediumDashDotDot"
  | "mediumDashed"
  | "none"
  | "slantDashDot"
  | "thick"
  | "thin";
// ...
export type XLSXHorizontalAlignment =
  | "general"
  | "left"
  | "center"
  | "right"
  | "fill"
  | "justify"
  | "centerContinuous"
  | "distributed";

export type XLSXVerticalAlignment = "top" | "center" | "bottom" | "justify" | "distributed";
// ...
export interface XLSXColor {
  auto?: boolean;
  indexed?: number;
  rgb?: string;
  tint?: number;
}